+++ src/components/TextForm.js
@@ -2,11 +2,16 @@ import React from "react";
 import { Button, Input } from "../elements";
 
 const TextForm = (props) => {
-  const { value, placeholder, btnName, _onSubmit } = props;
+  const { value, placeholder, btnName, _onSubmit, _onChange } = props;
 
   return (
     <form onSubmit={_onSubmit}>
-      <Input type="text" placeholder={placeholder} value={value} />
+      <Input
+        type="text"
+        placeholder={placeholder}
+        value={value}
+        _onChange={_onChange}
+      />
       <Button text={btnName} />
     </form>
   );
@@ -14,6 +19,7 @@ const TextForm = (props) => {
 
 TextForm.defaultProps = {
   _onSubmit: () => {},
+  _onChange: () => {},
   value: "",
   placeholder: "",
   btnName: "제출",
